Slim nvisy-core; dedup tabular/text; move retention resolver
- policy::redaction::TabularRedaction collapses to Cell(TextRedaction)
  + DropRow + DropColumn (was 9 duplicated variants). Engine gains
  anonymizer/text_op.rs owning the TextRedaction -> elide operator
  bridge; text.rs and tabular.rs share it.
- policy::retention split: RetentionScope moves to scope.rs,
  RetentionPolicy stays in mod.rs. resolve_retention() moves to
  nvisy_engine::retention::resolve (with its 8 tests) — core stays
  a frozen data model, resolution strategy lives with its callers.
- source.rs deleted. ContentSource was a newtype over Uuid with
  unused parent_id lineage plumbing; five context/ field sites
  swap to Uuid directly.
- file.rs promoted to file/ folder module: metadata.rs (FileMetadata),
  lineage.rs (FileLineage), raw.rs (RawDocument).
