+++ crates/nvisy-core/src/file/metadata.rs
@@ -1,29 +1,14 @@
-//! [`FileMetadata`]: persisted descriptor for an uploaded file.
-//!
-//! A file in the engine is a `(metadata, bytes)` pair: the bytes
-//! live in a blob-separated keyspace, the metadata in a small
-//! JSON keyspace. The split lets `list_files` enumerate every
-//! file for an actor without paying the cost of loading the
-//! bytes. Both keyspaces key by `(actor_id, file_id)`.
-//!
-//! The descriptor mirrors what [`DocumentInput`] carries on a
-//! [`StartBatch`] — the same `descriptor_labels` /
-//! `descriptor_metadata` gate policies via
-//! [`DocumentPredicate`]. When a run references a stored file,
-//! the run inherits these gates.
-//!
-//! [`DocumentInput`]: https://docs.rs/nvisy-engine/latest/nvisy_engine/runs/struct.DocumentInput.html
-//! [`StartBatch`]: https://docs.rs/nvisy-engine/latest/nvisy_engine/runs/struct.StartBatch.html
-//! [`DocumentPredicate`]: crate::policy::DocumentPredicate
+//! [`FileMetadata`]: the persisted descriptor for one stored file.
 
 use std::collections::HashMap;
 
-use bytes::Bytes;
 use hipstr::HipStr;
 use jiff::Timestamp;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
+
+use super::FileLineage;
 
 /// Descriptor for one stored file. Persisted as a small JSON
 /// blob next to the raw bytes; clients consume this through
@@ -82,44 +67,4 @@ pub struct FileMetadata {
     /// [`DocumentPredicate::HasMetadata`]: crate::policy::DocumentPredicate::HasMetadata
     #[serde(default, skip_serializing_if = "HashMap::is_empty")]
     pub descriptor_metadata: HashMap<String, String>,
-}
-
-/// A document ready for the codec: raw bytes plus the hints the
-/// codec needs to resolve a decoder.
-///
-/// In-memory carrier — not persisted, not on the wire. Built from
-/// a [`FileMetadata`] + its bytes when the engine reads a stored
-/// file, or from an upload's body + headers at ingest.
-#[derive(Debug, Clone)]
-pub struct RawDocument {
-    /// Raw file bytes.
-    pub bytes: Bytes,
-    /// File extension the codec registry resolves on (e.g.
-    /// `"txt"`, `"pdf"`, `"png"`). Case-insensitive, no leading
-    /// dot.
-    pub extension: HipStr<'static>,
-    /// Caller-supplied MIME hint (e.g. `application/pdf`). The
-    /// codec uses [`extension`], not this; recorded for audit and
-    /// for clients that round-trip metadata.
-    ///
-    /// [`extension`]: RawDocument::extension
-    pub content_type: Option<HipStr<'static>>,
-}
-
-/// Provenance for a [`FileMetadata`]. Uploaded files carry
-/// `None`; engine-produced files (today: redaction apply
-/// outputs) carry one of these variants so audits and clients
-/// can trace any file back to what produced it.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-#[serde(tag = "kind", rename_all = "camelCase")]
-pub enum FileLineage {
-    /// Output of a redaction apply. `runId` is the run that
-    /// produced this file; `sourceFileId` is the input file the
-    /// run read.
-    RedactedFrom {
-        /// Run that produced this file (`/redactions/{runId}`).
-        run_id: Uuid,
-        /// Original input file the run read.
-        source_file_id: Uuid,
-    },
 }
