+++ crates/nvisy-core/src/file/mod.rs
@@ -1,0 +1,27 @@
+//! File types: [`FileMetadata`] (the persisted descriptor),
+//! [`FileLineage`] (provenance for engine-produced files), and
+//! [`RawDocument`] (the in-memory codec input).
+//!
+//! A file in the engine is a `(metadata, bytes)` pair: the bytes
+//! live in a blob-separated keyspace, the metadata in a small
+//! JSON keyspace. The split lets `list_files` enumerate every
+//! file for an actor without paying the cost of loading the
+//! bytes. Both keyspaces key by `(actor_id, file_id)`.
+//!
+//! The descriptor mirrors what [`DocumentInput`] carries on a
+//! [`StartBatch`] — the same `descriptor_labels` /
+//! `descriptor_metadata` gate policies via
+//! [`DocumentPredicate`]. When a run references a stored file,
+//! the run inherits these gates.
+//!
+//! [`DocumentInput`]: https://docs.rs/nvisy-engine/latest/nvisy_engine/runs/struct.DocumentInput.html
+//! [`StartBatch`]: https://docs.rs/nvisy-engine/latest/nvisy_engine/runs/struct.StartBatch.html
+//! [`DocumentPredicate`]: crate::policy::DocumentPredicate
+
+mod lineage;
+mod metadata;
+mod raw;
+
+pub use self::lineage::FileLineage;
+pub use self::metadata::FileMetadata;
+pub use self::raw::RawDocument;
