+++ crates/nvisy-core/src/lib.rs
@@ -7,8 +7,6 @@ pub mod file;
 pub mod plan;
 pub mod policy;
 pub mod service;
-pub mod source;
 
 pub use self::file::{FileLineage, FileMetadata, RawDocument};
 pub use self::service::{Error, ErrorKind, Result};
-pub use self::source::ContentSource;
