+++ crates/nvisy-engine/src/engine/anonymizer/dispatch.rs
@@ -163,4 +163,3 @@ where
         redactions,
     )
 }
-
