+++ crates/nvisy-engine/src/engine/mod.rs
@@ -75,10 +75,9 @@ use elide_core::modality::audio::Audio;
 use elide_core::modality::image::Image;
 use elide_core::modality::tabular::Tabular;
 use elide_core::modality::text::Text;
-use nvisy_core::RawDocument;
 use nvisy_core::plan::AnalyzerParams;
 use nvisy_core::policy::{Policy, RuleAction};
-use nvisy_core::{Error, Result};
+use nvisy_core::{Error, RawDocument, Result};
 use uuid::Uuid;
 
 use self::report::{
